restructure utils in src
