+++ src/ClimaAtmos.jl
@@ -5,12 +5,12 @@ using NVTX, Colors
 include("Parameters.jl")
 import .Parameters as CAP
 
-include("common_spaces.jl")
-include("types.jl")
-include("utilities.jl")
-include("examples_utilities.jl")
-include("classify_case.jl")
-include("topography_helper.jl")
+include(joinpath("utils", "common_spaces.jl"))
+include(joinpath("utils", "types.jl"))
+include(joinpath("utils", "utilities.jl"))
+include(joinpath("utils", "debug_utils.jl"))
+include(joinpath("utils", "classify_case.jl"))
+include(joinpath("utils", "topography_helper.jl"))
 
 include(joinpath("parameterizations", "radiation", "radiation_utilities.jl"))
 include(joinpath("parameterizations", "radiation", "RRTMGPInterface.jl"))
@@ -20,13 +20,13 @@ include(joinpath("parameterizations", "radiation", "radiation.jl"))
 include("TurbulenceConvection/TurbulenceConvection.jl")
 import .TurbulenceConvection as TC
 
-include("thermo_state.jl")
+include(joinpath("utils", "thermo_state.jl"))
 include("precomputed_quantities.jl")
 
-include("callbacks.jl")
+include(joinpath("callbacks", "callbacks.jl"))
 
 include(joinpath("InitialConditions", "InitialConditions.jl"))
-include("discrete_hydrostatic_balance.jl")
+include(joinpath("utils", "discrete_hydrostatic_balance.jl"))
 
 include(joinpath("tendencies", "implicit", "wfact.jl"))
 include(joinpath("tendencies", "implicit", "schur_complement_W.jl"))
@@ -60,7 +60,7 @@ include(joinpath("tendencies", "advection.jl"))
 
 include("staggered_nonhydrostatic_model.jl")
 
-include("model_getters.jl") # high-level (using parsed_args) model getters
-include("examples_types.jl")
+include(joinpath("utils", "model_getters.jl")) # high-level (using parsed_args) model getters
+include(joinpath("utils", "type_getters.jl"))
 
 end # module
